requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
google-generativeai==0.3.2
python-dotenv==1.0.0
gunicorn
